remove dashboard
