+++ app/assets/stylesheets/components.css
@@ -14,12 +14,6 @@
   padding: 5px 10px;
 }
 
-.total-budget {
-  font-size: 60px;
-  color: green;
-  text-align: center;
-}
-
 .merch_period_picker {
   padding-left: 7.5px;
   padding-right: 7.5px;
