--- app/assets/stylesheets/dashboard.css
@@ -1,8 +1,0 @@
-#style-selection-bar {
-  background: #efefef;
-}
-
-hr {
-  margin-top: 0px;
-  margin-bottom: 0px;
-}
