--- app/assets/stylesheets/reminders.scss
@@ -1,29 +1,0 @@
-// Place all the styles related to the Reminders controller here.
-// They will automatically be included in application.css.
-// You can use Sass (SCSS) here: http://sass-lang.com/
-
-.reminders {
-  list-style: none;
-  padding: 0;
-  li {
-    padding: 10px 0;
-    border-top: 1px solid #e8e8e8;
-  }
-  .content {
-    display: block;
-    margin-left: 10px;
-  }
-  .timestamp {
-    display: block;
-    margin-left: 10px;
-  }
-
-  .disabled{
-    background-color: #ccc;
-  }
-}
-
-.pagination {
-  display: block;
-  text-align: center;
-}
