--- app/views/dashboard/budgets/edit.js.erb
@@ -1,1 +1,0 @@
-$('#budget-card .user-input').html("<%= j(render partial: 'edit') %> ");
