--- app/views/dashboard/cards/_budget_breakdown.html.haml
@@ -1,53 +1,0 @@
-.card
-  .row
-    .col-md-12
-      %canvas#pie-chart{style:'width: 200px;height:150px;'}
-  .row
-    .col-md-4{data:{placement: "bottom"},"data-toggle": "tooltip", title: "Open = Total Budget - Placed"}
-      %small.text-muted
-        Open
-        %br
-      = number_to_currency(@budget[:open] + @budget[:pending])
-    .col-md-4{data:{placement: "bottom"},"data-toggle": "tooltip", title: "Pending is a subset of Open"}
-      %small.text-muted
-        Pending
-        %br
-      = number_to_currency(@budget[:pending])
-    .col-md-4{data:{placement: "bottom"},"data-toggle": "tooltip", title: "Placed funds are no longer available"}
-      %small.text-muted
-        Placed
-        %br
-      = number_to_currency(@budget[:placed])
-
-
-:javascript
-  $(function(){
-    var pie_data = {
-        labels: [
-            "Open",
-            "Pending",
-            "Placed"
-        ],
-        datasets: [
-            {
-                data: [#{@budget[:open]}, #{@budget[:pending]}, #{@budget[:placed]}],
-                backgroundColor: [
-                    "#36A2EB",
-                    "#36A2EB",
-                    "#eb7f36"
-                ],
-                hoverBackgroundColor: [
-                    "#36A2EB",
-                    "#36A2EB",
-                    "#eb7f36"
-                ]
-            }]
-    };
-
-    var pie_ctx = document.getElementById("pie-chart");
-    var myPieChart = new Chart(pie_ctx,{
-        type: 'pie',
-        data: pie_data,
-        options: {}
-    });
-  })
