--- app/views/dashboard/cards/_fetch_reminders.html.haml
@@ -1,8 +1,0 @@
-.card.card-row
-  %i.fa.fa-pencil.fa-spin
-  fetching reminders
-  
-:javascript
-  $(function(){
-    $.get("#{reminders_path}");
-  });
