--- app/views/dashboard/cards/_incomplete_appointments.html.haml
@@ -1,9 +1,0 @@
-.card
-  %table{style: 'width:100%'}
-    %tr{data:{toggle: "tooltip"}, title: "Incomplete Appointments contain Incomplete Styles"}
-      %td
-        %small.text-muted
-          Incomplete Appointments:
-      %td
-        %h3.no-margin.text-right
-          =@dashboard_stats[:incomplete_appointments].count
