--- app/views/dashboard/cards/_incomplete_styles.html.haml
@@ -1,9 +1,0 @@
-.card.card-row
-  %table{style: 'width:100%'}
-    %tr{data:{placement: "bottom"},"data-toggle": "tooltip", title: "Incomplete Styles are Created or Pending"}
-      %td
-        %small.text-muted
-          Incomplete Styles
-      %td
-        %h3.no-margin.text-right
-          = @dashboard_stats[:incomplete_styles].count
