--- app/views/dashboard/cards/_month_picker.html.haml
@@ -1,26 +1,0 @@
-.card.card-row
-  .form-group
-    %small.text-muted
-      Merch Month Period
-  = form_tag root_path, method: :get do
-    .merch_period_row.row.form-group
-      .merch_period_picker.col-md-4
-        %label Year
-        %select.form-control{name: 'merch_year'}
-          = options_for_merch_year_picker(params)
-      .merch_period_picker.col-md-5
-        %label Merch Month
-        %select.form-control{name: 'merch_month'}
-          = options_for_merch_month_picker(params)
-      .merch_period_picker.col-md-3
-        %label &nbsp
-        %button.btn.btn-default.form-control{type: "submit"} >
-  .row.merch_period_row
-    .col-md-6
-      %h5 Period Start:
-      .mark
-        = start_of_period(params)
-    .col-md-6
-      %h5 Period End:
-      .mark
-        = end_of_period(params)
