--- app/views/dashboard/cards/_reminders.html.haml
@@ -1,28 +1,0 @@
-.card.card-row
-  %small.text-muted
-    = "Reminders"
-  .row
-    = form_tag reminders_path, method: :post, remote: true do
-      .col-md-12
-        .input-group
-          = text_field_tag 'reminder[content]', '', class: 'form-control'
-          %span.input-group-btn
-            = submit_tag '+', class: 'btn btn-primary'
-  %ol.reminders
-    - @reminders.each do |r|
-      %li{id: "reminder-#{r.id}", class: ('disabled' if r.completed?) }
-        %span.content.small
-          = form_tag reminder_path(r), method: :put, style: "display: inline;", remote: true do
-            = hidden_field_tag 'reminder[completed]', false
-            = check_box_tag 'reminder[completed]', true, r.completed, onclick: "$(this).parents('form').submit()"
-          = r.content
-        %span.timestamp.small.text-muted
-          Created #{time_ago_in_words(r.created_at)} ago.
-          = link_to edit_reminder_path(r), remote: true do
-            %i.fa.fa-pencil
-          = link_to reminder_path(r), method: :delete, remote: true do
-            %i.fa.fa-trash
-    = js_will_paginate @reminders
-
-    -if @reminders.blank?
-      No reminders yet.
