--- app/views/dashboard/cards/_total_budget.html.haml
@@ -1,12 +1,0 @@
-#budget-card.card
-  %small.text-muted
-    =link_to 'Total Budget', budgets_path
-  .total-budget
-    = number_to_currency(@budget[:total])
-  %small.text-muted
-    - if @budget[:budget].nil?
-      = link_to "Add Budget", new_dashboard_budget_path(params), remote: :true
-    - else
-      = link_to "Edit Budget", edit_dashboard_budget_path(@budget[:current_budget_id], params), remote: true
-  .user-input
-    - # this is a container div
