--- app/views/dashboard/show.html.haml
@@ -1,68 +1,0 @@
-.row.card-row
-  .col-sm-4
-    = render 'dashboard/cards/total_budget'
-    = render 'dashboard/cards/month_picker'
-  .col-sm-4
-    = render 'dashboard/cards/budget_breakdown'
-  .col-sm-4
-    = render 'dashboard/cards/incomplete_appointments'
-    = render 'dashboard/cards/incomplete_styles'
-    #reminders-container
-      = render 'dashboard/cards/fetch_reminders'
-
-%hr
-.row.card-row.no-margin
-  .col-md-6
-    %h3.text-muted
-      Appointments:
-    #style-tabs
-      %ul.nav.nav-tabs
-        %li.active{data:{toggle: "tooltip"}, title: "Appointments with styles that have due dates in the specified month"}
-          %a{"data-toggle" => "tab", :href => "#appointments"}
-            Appointments
-        %li{data:{toggle: "tooltip"}, title: "Appointments with styles that have no specified due date"}
-          %a{"data-toggle" => "tab", href: "#orphan-appointments"}
-            Orphan Appointments
-      .tab-content
-        .tab-pane.active#appointments
-          %ul.list-group
-            - @dashboard_stats[:incomplete_appointments].each_with_index do |appointment|
-              .list-group-item
-                = link_to appointment.name, appointment_path(appointment)
-                %i.fa.fa-exclamation.text-danger{data:{toggle: "tooltip"}, title: "This appointment has some styles that are still created or pending."}
-
-            - @dashboard_stats[:complete_appointments].each do |appointment|
-              .list-group-item
-                = link_to appointment.name, appointment_path(appointment)
-                %i.fa.fa-check.text-success{data:{toggle: "tooltip"}, title: "All styles have been processed!"}
-        .tab-pane#orphan-appointments
-          %ul.list-group
-            - @dashboard_stats[:dateless_style_appointments].each do |appointment|
-              .list-group-item
-                = link_to appointment.name, appointment_path(appointment)
-
-  .col-md-6
-    %h3.text-muted
-      Styles
-    #style-tabs
-      %ul.nav.nav-tabs
-        %li.active
-          %a{"data-toggle" => "tab", :href => "#created-styles"}
-            Created
-            (#{@dashboard_stats[:created_styles].count})
-        %li
-          %a{"data-toggle" => "tab", :href => "#pending-styles"}
-            Pending
-            (#{@dashboard_stats[:pending_styles].count})
-        %li
-          %a{"data-toggle" => "tab", :href => "#placed-styles"}
-            Placed
-            (#{@dashboard_stats[:placed_styles].count})
-      .tab-content
-        - %w(created pending placed).each_with_index do |status, i|
-          .tab-pane{id: "#{status}-styles", class: ('active' if i == 0).to_s }
-            %ul.list-group
-              - @dashboard_stats["#{status}_styles".to_sym].each do |style|
-                .list-group-item
-                  = link_to style.vendor_style_number, appointment_style_path(style.appointment.id, style)
-                  = style.status
