--- app/views/reminders/_edit.html.haml
@@ -1,4 +1,0 @@
-= form_for @reminder, url:reminder_path(@reminder) , method: :put, remote: true do |f|
-  .form-inline
-    =f.text_field :content, class: 'form-control'
-    =f.submit 'Save', class: 'btn btn-sm btn-primary'
