--- app/views/reminders/edit.js.erb
@@ -1,1 +1,0 @@
-$('#reminder-<%= @reminder.id %> .content').html("<%= j(render partial: 'reminders/edit') %> ")
