--- app/views/reminders/index.js.erb
@@ -1,1 +1,0 @@
-$('#reminders-container').html("<%= j(render partial: 'dashboard/cards/reminders') %> ")
