--- spec/controllers/reminders_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe RemindersController, type: :controller do
-
-end
