--- spec/factories/reminders.rb
@@ -1,6 +1,0 @@
-FactoryGirl.define do
-  factory :reminder do
-    content "MyText"
-    user nil
-  end
-end
