--- spec/models/reminder_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe Reminder, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
